chore: Upgrade springboot to 2.7.4 This requires us to upgrade spring-cloud to 2.1.4 and remove a circular dependency from our @Configuration beans
