+++ build.gradle
@@ -1,6 +1,6 @@
 buildscript {
     ext {
-        springBootVersion = "2.5.12"
+        springBootVersion = "2.7.4"
         httpclientVersion = "4.5.13"
         aerospikeClientVersion = findProperty("aerospikeClientVersion") ?: "6.1.2"
     }
@@ -63,7 +63,7 @@ dependencies {
     implementation("org.msgpack:jackson-dataformat-msgpack:0.9.3")
     implementation("org.springframework.retry:spring-retry:1.3.4")
     implementation("org.springframework:spring-aspects:5.3.23")
-    implementation("org.springframework.cloud:spring-cloud-starter-circuitbreaker-resilience4j:2.0.2")
+    implementation("org.springframework.cloud:spring-cloud-starter-circuitbreaker-resilience4j:2.1.4")
     implementation("org.springdoc:springdoc-openapi-ui:1.6.12")
     implementation("javax.inject:javax.inject:1")
     implementation("com.google.guava:guava:31.0.1-jre")
